fix: widen *_cents columns to bigInteger and add stock CHECK constraint
Money columns (sale_price_cents, average_cost_cents, total_cents,
profit_cents, subtotal_cents, unit_price_cents,
average_cost_snapshot_cents, item_profit_cents) were 32-bit
integer/unsignedInteger, overflowing beyond ~R$21.4M (2^31-1 cents).
Widen all of them to bigInteger/unsignedBigInteger, signed where a
negative value is valid (profit_cents, item_profit_cents,
average_cost_cents, average_cost_snapshot_cents) and unsigned
elsewhere. No production data exists yet, so migration files are
edited in place rather than adding alter-table migrations.

Also adds the missing DB-level CHECK constraint
(current_stock >= 0) on products, matching the pattern already used
for the other three constrained columns in this codebase.

+++ database/migrations/2026_08_05_162951_create_purchases_table.php
@@ -11,7 +11,7 @@ return new class extends Migration
         Schema::create('purchases', function (Blueprint $table) {
             $table->id();
             $table->string('supplier');
-            $table->integer('total_cents');
+            $table->unsignedBigInteger('total_cents');
             $table->timestamps();
         });
     }
